Merge branch rec_pipes into master
Merge + refactor pipes to manage special case with built-ins (ie. `exit
| echo` which doesnt exit). Built-ins are now always run into a
subprocess with fork(). We now check if the number of pipes is 0 to see
if we need to run exit() in the main process if called.

Also refactor libft.h to remove unnecessary includes. This change
propagated to a few other C files in the main project which needed some
includes added.

Fixed a few norme cases.